♻️ refactor :  RecipeDataType 분리
